sync_release-manifest: remove top-level backstage directory
The top-level backstage/ directory included in
this repo means we can't use the yarn-install
backstage action (or we'd have to add a `path`
input option to it).

Looking at this workflow, I can't actually spot
why the top-level `backstage` dir is needed -
nothing is ever written outside it as far as I can
spot. With that in mind, proposing we remove it to
simplify the workflow and enable the use of the
`yarn-install` action without needing to extend it.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>